Relocate some classes to types package
