+++ api/src/main/java/de/verdox/mccreativelab/wrapper/entity/MCCAttributeInstance.java
@@ -4,7 +4,7 @@ import de.verdox.mccreativelab.wrapper.annotations.MCCInstantiationSource;
 import de.verdox.mccreativelab.wrapper.annotations.MCCLogic;
 import de.verdox.mccreativelab.wrapper.item.MCCAttributeModifier;
 import de.verdox.mccreativelab.wrapper.registry.MCCReference;
-import de.verdox.mccreativelab.wrapper.world.chunk.MCCChunkSection;
+import de.verdox.mccreativelab.wrapper.types.MCCAttribute;
 import net.kyori.adventure.key.Key;
 import org.jetbrains.annotations.NotNull;
