+++ api/src/main/java/de/verdox/mccreativelab/wrapper/entity/MCCAttributeMap.java
@@ -1,6 +1,7 @@
 package de.verdox.mccreativelab.wrapper.entity;
 
 import de.verdox.mccreativelab.wrapper.registry.MCCReference;
+import de.verdox.mccreativelab.wrapper.types.MCCAttribute;
 import net.kyori.adventure.key.Key;
 
 /**
