+++ api/src/main/java/de/verdox/mccreativelab/wrapper/types/MCCJukeboxSong.java
@@ -1,12 +1,8 @@
 package de.verdox.mccreativelab.wrapper.types;
-
-import net.kyori.adventure.sound.Sound;
 import de.verdox.mccreativelab.wrapper.registry.MCCReference;
+import net.kyori.adventure.sound.Sound;
 import net.kyori.adventure.text.Component;
-import de.verdox.mccreativelab.wrapper.MCCWrapped;
-
-public interface MCCJukeboxSong extends MCCWrapped  {
-
+public interface MCCJukeboxSong  {
 	public MCCReference<Sound> getSoundEvent();
 
 	public Component getDescription();
@@ -14,5 +10,4 @@ public interface MCCJukeboxSong extends MCCWrapped  {
 	public float getLengthInSeconds();
 
 	public int getComparatorOutput();
-
 }
