+++ api/src/main/java/de/verdox/mccreativelab/wrapper/types/MCCPoiType.java
@@ -1,15 +1,12 @@
 package de.verdox.mccreativelab.wrapper.types;
+import java.util.Set;
 
 import de.verdox.mccreativelab.wrapper.block.MCCBlockState;
-import java.util.Set;
-import de.verdox.mccreativelab.wrapper.MCCWrapped;
 
-public interface MCCPoiType extends MCCWrapped  {
-
+public interface MCCPoiType  {
 	public Set<MCCBlockState> getMatchingStates();
 
 	public int getMaxTickets();
 
 	public int getValidRange();
-
 }
