+++ api/src/main/java/de/verdox/mccreativelab/wrapper/types/MCCTrimPattern.java
@@ -1,13 +1,9 @@
 package de.verdox.mccreativelab.wrapper.types;
-
 import de.verdox.mccreativelab.wrapper.registry.MCCReference;
 import net.kyori.adventure.text.Component;
-import de.verdox.mccreativelab.wrapper.MCCWrapped;
 import net.kyori.adventure.key.Key;
 import de.verdox.mccreativelab.wrapper.item.MCCItemType;
-
-public interface MCCTrimPattern extends MCCWrapped  {
-
+public interface MCCTrimPattern  {
 	public Key getAssetId();
 
 	public MCCReference<MCCItemType> getTemplateItem();
@@ -15,5 +11,4 @@ public interface MCCTrimPattern extends MCCWrapped  {
 	public Component getDescription();
 
 	public boolean getDecal();
-
 }
